apps/sketch2bim/infra/hetzner/requirements.txt
# Hetzner Cloud Infrastructure Requirements
hcloud>=1.19.0
python-dotenv>=1.0.0

